Added config.rs
