+++ bin/node/src/commands/mod.rs
@@ -1,5 +1,3 @@
 mod genesis;
+pub mod start;
 pub use genesis::make_genesis;
-
-mod start;
-pub use start::{start_block_producer, start_node, start_rpc, start_store, StartCommandConfig};
