+++ bin/node/src/commands/start.rs
@@ -4,24 +4,14 @@ use anyhow::Result;
 use miden_node_block_producer::{config::BlockProducerConfig, server as block_producer_server};
 use miden_node_rpc::{config::RpcConfig, server as rpc_server};
 use miden_node_store::{config::StoreConfig, db::Db, server as store_server};
-use serde::{Deserialize, Serialize};
 use tokio::task::JoinSet;
 
-// Config
-// ================================================================================================
-
-/// Node configuration.
-#[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Serialize, Deserialize)]
-pub struct StartCommandConfig {
-    pub block_producer: BlockProducerConfig,
-    pub rpc: RpcConfig,
-    pub store: StoreConfig,
-}
+use crate::config::NodeConfig;
 
 // START
 // ===================================================================================================
 
-pub async fn start_node(config: StartCommandConfig) -> Result<()> {
+pub async fn start_node(config: NodeConfig) -> Result<()> {
     let mut join_set = JoinSet::new();
     let db = Db::setup(config.store.clone()).await?;
     join_set.spawn(store_server::serve(config.store, db));
@@ -61,78 +51,4 @@ pub async fn start_store(config: StoreConfig) -> Result<()> {
     store_server::serve(config, db).await?;
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::path::PathBuf;
-
-    use figment::Jail;
-    use miden_node_block_producer::config::BlockProducerConfig;
-    use miden_node_rpc::config::RpcConfig;
-    use miden_node_store::config::StoreConfig;
-    use miden_node_utils::config::{load_config, Endpoint};
-
-    use super::StartCommandConfig;
-    use crate::NODE_CONFIG_FILE_PATH;
-
-    #[test]
-    fn test_node_config() {
-        Jail::expect_with(|jail| {
-            jail.create_file(
-                NODE_CONFIG_FILE_PATH,
-                r#"
-                    [block_producer]
-                    endpoint = { host = "127.0.0.1",  port = 8080 }
-                    store_url = "http://store:8000"
-                    verify_tx_proofs = true
-
-                    [rpc]
-                    endpoint = { host = "127.0.0.1",  port = 8080 }
-                    store_url = "http://store:8000"
-                    block_producer_url = "http://block_producer:8001"
-
-                    [store]
-                    endpoint = { host = "127.0.0.1",  port = 8080 }
-                    database_filepath = "local.sqlite3"
-                    genesis_filepath = "genesis.dat"
-                "#,
-            )?;
-
-            let config: StartCommandConfig =
-                load_config(PathBuf::from(NODE_CONFIG_FILE_PATH).as_path()).extract()?;
-
-            assert_eq!(
-                config,
-                StartCommandConfig {
-                    block_producer: BlockProducerConfig {
-                        endpoint: Endpoint {
-                            host: "127.0.0.1".to_string(),
-                            port: 8080,
-                        },
-                        store_url: "http://store:8000".to_string(),
-                        verify_tx_proofs: true
-                    },
-                    rpc: RpcConfig {
-                        endpoint: Endpoint {
-                            host: "127.0.0.1".to_string(),
-                            port: 8080,
-                        },
-                        store_url: "http://store:8000".to_string(),
-                        block_producer_url: "http://block_producer:8001".to_string(),
-                    },
-                    store: StoreConfig {
-                        endpoint: Endpoint {
-                            host: "127.0.0.1".to_string(),
-                            port: 8080,
-                        },
-                        database_filepath: "local.sqlite3".into(),
-                        genesis_filepath: "genesis.dat".into()
-                    },
-                }
-            );
-
-            Ok(())
-        });
-    }
 }
